I added 2 more pages.
One page will be for a starter table with one column that is a string, and one page will be for a many to many table.

I have added 2 buttons on the main page to go to these pages.

I have an error. I need to add an event. Once I add an event, the error will go away. There was an error with span, and I fixed it with margin, and then that error went away, and then I ran it and thought there were going to be no errors, but I got a build error.